Make floats double precision
Merge PR from @jmchandonia converting float types to double precision.

Follow-up PR will address:
- Python 3.9-3.11 compatibility issues
- Copilot review feedback
- Test coverage

Original PR: #33

+++ src/linkml_coral/convert/convert_to_linkml.py
@@ -55,7 +55,7 @@ class TypedefToLinkMLConverter:
         type_mapping = {
             "text": "string",
             "int": "integer", 
-            "float": "float",
+            "float": "double",
             "term": "string",  # Terms will be validated with enums
             "[text]": "string",  # Array types handled with multivalued
             "[ref]": "string"   # References handled with ranges
